# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/properties": {
            "get": {
                "tags": [
                    "Property"
                ],
                "summary": "Get a listing of the Properties.",
                "description": "Get all Properties",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Property"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Property"
                ],
                "summary": "Store a newly created Property in storage",
                "description": "Store Property",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Property that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Property"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Property"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/properties/{id}": {
            "get": {
                "tags": [
                    "Property"
                ],
                "summary": "Display the specified Property",
                "description": "Get Property",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Property",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Property"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Property"
                ],
                "summary": "Update the specified Property in storage",
                "description": "Update Property",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Property",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Property that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Property"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Property"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Property"
                ],
                "summary": "Remove the specified Property from storage",
                "description": "Delete Property",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Property",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Property": {
            "required": [
                "status",
                "property_age",
                "property_categorie_id",
                "user_id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "address": {
                    "description": "address",
                    "type": "string"
                },
                "lat": {
                    "description": "lat",
                    "type": "number",
                    "format": "number"
                },
                "lng": {
                    "description": "lng",
                    "type": "number",
                    "format": "number"
                },
                "status": {
                    "description": "status",
                    "type": "integer",
                    "format": "int32"
                },
                "room_number": {
                    "description": "room_number",
                    "type": "integer",
                    "format": "int32"
                },
                "property_age": {
                    "description": "property_age",
                    "type": "integer",
                    "format": "int32"
                },
                "furnished": {
                    "description": "furnished",
                    "type": "integer",
                    "format": "int32"
                },
                "air_conditioner": {
                    "description": "air_conditioner",
                    "type": "integer",
                    "format": "int32"
                },
                "space": {
                    "description": "space",
                    "type": "integer",
                    "format": "int32"
                },
                "price": {
                    "description": "price",
                    "type": "number",
                    "format": "number"
                },
                "note": {
                    "description": "note",
                    "type": "string"
                },
                "destination": {
                    "description": "destination",
                    "type": "string"
                },
                "advertiser_status": {
                    "description": "advertiser_status",
                    "type": "string"
                },
                "land_number": {
                    "description": "land_number",
                    "type": "number"
                },
                "floor": {
                    "description": "floor",
                    "type": "number"
                },
                "the_number_apartments": {
                    "description": "the_number_apartments",
                    "type": "number"
                },
                "the_number_stores": {
                    "description": "the_number_stores",
                    "type": "number"
                },
                "property_categorie_id": {
                    "description": "property_categorie_id",
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "description": "user_id",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}